backend/requirements.txt
Flask==3.1.1
flask_cors==5.0.1
groq==0.25.0
python-dotenv==1.1.0
